requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
httpx==0.27.0
groq==0.9.0
supabase==2.4.6
cloudinary==1.40.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
apscheduler==3.10.4
firebase-admin==6.5.0
Pillow==10.3.0
aiofiles==23.2.1
pydantic==2.7.1
pydantic-settings==2.2.1
slowapi==0.1.9
loguru==0.7.2
